Rename `service/engine.rs` -> `service/syncing_service.rs`
