+++ substrate/client/network/sync/src/lib.rs
@@ -18,7 +18,7 @@
 
 //! Blockchain syncing implementation in Substrate.
 
-pub use service::engine::SyncingService;
+pub use service::syncing_service::SyncingService;
 pub use types::{SyncEvent, SyncEventStream, SyncState, SyncStatus, SyncStatusProvider};
 
 mod block_announce_validator;
