+++ substrate/client/network/sync/src/service/mod.rs
@@ -18,6 +18,6 @@
 
 //! `SyncingEngine`-related service code
 
-pub mod engine;
 pub mod mock;
 pub mod network;
+pub mod syncing_service;
